Small Change
Made small changes to character menu and HUD

+++ Assets/Scripts/HUD.cs
@@ -15,7 +15,7 @@ public class HUD : MonoBehaviour
 
     private void Start()
     {
-        health = GameManager.instance.player.maxHitPoint;
+        health = (float) GameManager.instance.player.maxHitPoint;
         maxHealth = health;
     }
